Issue #9072 refactor jetty-ee module to remove it
