--- jetty-core/jetty-ee/src/main/java/module-info.java
@@ -1,26 +1,0 @@
-//
-// ========================================================================
-// Copyright (c) 1995 Mort Bay Consulting Pty Ltd and others.
-//
-// This program and the accompanying materials are made available under the
-// terms of the Eclipse Public License v. 2.0 which is available at
-// https://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
-// which is available at https://www.apache.org/licenses/LICENSE-2.0.
-//
-// SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
-// ========================================================================
-//
-
-module org.eclipse.jetty.ee
-{
-    requires org.slf4j;
-
-    requires transitive org.eclipse.jetty.io;
-    requires transitive org.eclipse.jetty.security;
-
-    // Only required if using JMX.
-    requires static org.eclipse.jetty.jmx;
-
-    exports org.eclipse.jetty.ee;
-    exports org.eclipse.jetty.ee.security;
-}
