+++ jetty-core/jetty-server/src/main/config/modules/server.mod
@@ -19,7 +19,6 @@ lib/jetty-server-${jetty.version}.jar
 lib/jetty-xml-${jetty.version}.jar
 lib/jetty-util-${jetty.version}.jar
 lib/jetty-io-${jetty.version}.jar
-lib/jetty-ee-${jetty.version}.jar
 
 [xml]
 etc/jetty.xml
