+++ jetty-core/pom.xml
@@ -17,7 +17,6 @@
     <module>jetty-client</module>
     <module>jetty-deploy</module>
     <module>jetty-demos</module>
-    <module>jetty-ee</module>
     <module>jetty-fcgi</module>
     <module>jetty-http</module>
     <module>jetty-http2</module>
