+++ jetty-ee9/jetty-ee9-nested/src/main/java/module-info.java
@@ -14,7 +14,7 @@
 module org.eclipse.jetty.ee9.nested
 {
     requires transitive jetty.servlet.api;
-    requires transitive org.eclipse.jetty.ee;
+    requires transitive org.eclipse.jetty.security;
     requires transitive org.eclipse.jetty.http;
     requires transitive org.eclipse.jetty.server;
     requires transitive org.eclipse.jetty.session;
